Fleshed out peers, blockchain, connections and server
Ok, so we haven't been pushing as much as we should. We did a big group
of updates to each of these files and build the base of them.

I've added comments to each. There are some bits that really need to
change.

we also need to look at the last_block stuff, it's not written right and
will bomb.